Extract Dashboard Component
